Deduplicate body table: single shared source in Software/Common
The ascent autopilot's AP_BODIES table duplicated the pre-existing
celestial-body table in the KerbalDisplayCommon library. Consolidate to
one canonical source.

- Relocate the BodyParams struct + _bodyTable[] + getBodyParams() into a
  new header-only Software/Common/body_params.h (header-only so it is
  shareable across the two independent Arduino sketch trees without a
  separate compilation unit). Two Kerbol/Duna values that exceeded float's
  exact-integer range are now explicit float literals (value-identical).
- KerbalDisplayCommon.h now includes the shared header; its .cpp no longer
  defines the table/lookups. The display API (BodyParams, getBodyParams)
  is unchanged, so InfoDisp / Annunciator are unaffected.
- Controller_Main includes the shared header; the autopilot drops its
  duplicate AP_BODIES/BodyProfile/apLookupBody and reads getBodyParams()
  directly, using lowSpace (atmosphere top), minSafe (terrain), and
  hasAtmo. Parking-orbit and min-safe-altitude are now derived from those
  fields, so the autopilot adds no body data of its own.

Docs updated. Module type-checks clean (gnu++17) against the shared table.
